+++ Sources/Runner/main.swift
@@ -1,8 +1,21 @@
 import Foundation
+import RunnerLib
 import Logger
 
 /// Version for showing in verbose mode
 let DangerVersion = "0.7.1"
+
+fileprivate func runCommand(_ command: DangerCommand, logger: Logger) throws {
+    switch command {
+    case .ci, .local, .pr:
+        let exitCode = try runDangerJSCommandToRunDangerSwift(command, logger: logger)
+        exit(exitCode)
+    case .edit:
+        try editDanger(logger: logger)
+    case .runner:
+        try getDSLData(logger: logger, runDanger)
+    }
+}
 
 let cliLength = ProcessInfo.processInfo.arguments.count
 do {
@@ -12,16 +25,16 @@ do {
     logger.debug("Launching Danger Swift (v\(DangerVersion))")
 
     if cliLength > 1 {
-        switch(CommandLine.arguments[1]) {
-        case "ci", "local", "pr":
-            let exitCode = try runDangerJSCommandToRunDangerSwift(CommandLine.arguments[1], logger: logger)
-            exit(exitCode)
-        case "edit":
-            try editDanger(logger: logger)
-        case "runner":
-            try getDSLData(logger: logger, runDanger)
-
-        default:
+        let command = DangerCommand(rawValue: CommandLine.arguments[1])
+        
+        guard !CommandLine.arguments.contains("--help") else {
+            HelpMessagePresenter.showHelpMessage(command: command, logger: logger)
+            exit(0)
+        }
+        
+        if command != nil {
+            try runCommand(command!, logger: logger)
+        } else {
             fatalError("Danger Swift does not support this argument, it only handles ci, local, pr & edit'")
         }
     } else {