+++ Tests/RunnerLibTests/ImportsFinderTests.swift
@@ -9,13 +9,6 @@ import XCTest
 import RunnerLib
 
 final class ImportsFinderTests: XCTestCase {
-    static var allTests = [
-        ("testItRetunsTheCorrectFilePathsWhenThePassedStringContainsImports", testItRetunsTheCorrectFilePathsWhenThePassedStringContainsImports),
-        ("testItRetunsAnEmptyListWhenThePassedStringDoesntContainImports",
-         testItRetunsAnEmptyListWhenThePassedStringDoesntContainImports),
-        ("testLinuxTestSuiteIncludesAllTests", testLinuxTestSuiteIncludesAllTests)
-    ]
-    
     var importsFinder: ImportsFinder!
     
     override func setUp() {
@@ -29,15 +22,6 @@ final class ImportsFinderTests: XCTestCase {
 
     func testItRetunsAnEmptyListWhenThePassedStringDoesntContainImports() {
         checkReturnsTheCorrectFilePaths(string: stringWithoutImports, expectedResult: [])
-    }
-    
-    func testLinuxTestSuiteIncludesAllTests() {
-        #if !os(Linux)
-        let thisClass = type(of: self)
-        let linuxCount = thisClass.allTests.count
-        let darwinCount = thisClass.defaultTestSuite.tests.count
-        XCTAssertEqual(linuxCount, darwinCount, "\(darwinCount - linuxCount) tests are missing from allTests")
-        #endif
     }
 }