+++ CHANGELOG.md
@@ -14,10 +14,11 @@
 ## Master
 
 * Danger Swift can request the DSL from Danger JS under the hood, and listens async for STDIN - [@orta][]
+*Add --help option support to danger-swift [#122](https://github.com/danger/danger-swift/pull/122) by [@f-meloni][]
 
 ## 0.7.1
 
-Move report logic on a separate file [#111](https://github.com/danger/danger-swift/pull/111) by [@f-meloni][]
+*Move report logic on a separate file [#111](https://github.com/danger/danger-swift/pull/111) by [@f-meloni][]
 
 ## 0.7.0